# bpf/analyze: refactor reachability analysis and unused tail call elim