--- a/pkg/bpf/analyze/bitmap.go
+++ b/pkg/bpf/analyze/bitmap.go
@@ -8,21 +8,21 @@ import (
 	"unsafe"
 )
 
-// bitmap is a bitmap used to track which blocks are reachable in the control
+// Bitmap is a Bitmap used to track which blocks are reachable in the control
 // flow graph.
-type bitmap []uint64
+type Bitmap []uint64
 
-const wordSize = uint64(unsafe.Alignof(bitmap(nil)[0]) * 8)
+const wordSize = uint64(unsafe.Alignof(Bitmap(nil)[0]) * 8)
 
-// newBitmap returns a bitmap capable of tracking at least n items. All bits are
+// NewBitmap returns a bitmap capable of tracking at least n items. All bits are
 // false by default.
-func newBitmap(n uint64) bitmap {
-	return make(bitmap, (n+(wordSize-1))/wordSize)
+func NewBitmap(n uint64) Bitmap {
+	return make(Bitmap, (n+(wordSize-1))/wordSize)
 }
 
-// set sets the bit at index i to the given value. If i is out of bounds, it
+// Set sets the bit at index i to the given value. If i is out of bounds, it
 // does nothing.
-func (b bitmap) set(i uint64, value bool) {
+func (b Bitmap) Set(i uint64, value bool) {
 	word, bit := i/wordSize, i%wordSize
 	if word >= uint64(len(b)) {
 		return
@@ -35,9 +35,9 @@ func (b bitmap) set(i uint64, value bool) {
 	}
 }
 
-// get returns the value of the bit at index i. If i is out of bounds, it
+// Get returns the value of the bit at index i. If i is out of bounds, it
 // returns false.
-func (b bitmap) get(i uint64) bool {
+func (b Bitmap) Get(i uint64) bool {
 	word, bit := i/wordSize, i%wordSize
 	if word >= uint64(len(b)) {
 		return false
@@ -46,8 +46,8 @@ func (b bitmap) get(i uint64) bool {
 	return b[word]&(1<<(bit)) != 0
 }
 
-// popcount returns the number of bits set to true in the bitmap.
-func (b bitmap) popcount() uint64 {
+// Popcount returns the number of bits set to true in the bitmap.
+func (b Bitmap) Popcount() uint64 {
 	var count int
 	for _, w := range b {
 		count += bits.OnesCount64(w)