--- a/pkg/logging/logfields/logfields.go
+++ b/pkg/logging/logfields/logfields.go
@@ -1686,8 +1686,6 @@ const (
 
 	Section = "section"
 
-	Instruction = "instruction"
-
 	Reference = "reference"
 
 	MapRenames = "mapRenames"