… pkg/bpf/analyze/bitmap_test.go
… pkg/bpf/analyze/blocks_test.go
… pkg/bpf/analyze/reachability_test.go
… pkg/bpf/collection.go
… pkg/bpf/reachability.go
… pkg/bpf/testdata/unreachable-tailcall.c
… pkg/bpf/testdata/unreachable-tailcall.o
… pkg/bpf/unused_maps.go
… pkg/bpf/unused_maps_test.go
… pkg/bpf/unused_tailcalls.go
… pkg/bpf/unused_tailcalls_test.go