Some readability and performance improvements that couldn't hold up the merge of the initial work.

In short:
- added a unified `BlockIterator` that supports iterating forwards and backwards, as well as cloning for starting temporary backtracking sessions from the current instruction
- added a `Reachable` type to hold block reachability information, wrapping a Blocks and insns
- (not shown in benchmarks) Reachability is now computed only once per CollectionSpec load, as opposed to twice before the PR (once for maps, again for tail calls)
- Computing reachability now allocates about 80% less memory since Blocks no longer needs to be copied
- General refactoring for readability

```
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/bpf/analyze
cpu: AMD Ryzen 7 3700X 8-Core Processor
                 │   old.txt   │              new.txt               │
                 │   sec/op    │   sec/op     vs base               │
ComputeBlocks-16   633.7µ ± 2%   622.1µ ± 2%   -1.83% (p=0.015 n=6)
Reachability-16    32.87µ ± 2%   27.67µ ± 1%  -15.84% (p=0.002 n=6)
geomean            144.3µ        131.2µ        -9.10%

                 │   old.txt    │               new.txt               │
                 │     B/op     │     B/op      vs base               │
ComputeBlocks-16   372.6Ki ± 0%   372.5Ki ± 0%   -0.02% (p=0.002 n=6)
Reachability-16     8528.0 ± 0%     352.0 ± 0%  -95.87% (p=0.002 n=6)
geomean            55.71Ki        11.32Ki       -79.69%

                 │   old.txt   │              new.txt               │
                 │  allocs/op  │  allocs/op   vs base               │
ComputeBlocks-16   8.054k ± 0%   8.053k ± 0%   -0.01% (p=0.002 n=6)
Reachability-16     4.000 ± 0%    3.000 ± 0%  -25.00% (p=0.002 n=6)
geomean             179.5         155.4       -13.40%
```

Fixes: https://github.com/cilium/cilium/issues/41628